snow-rand {snow}	R Documentation

Uniform Random Number Generation in SNOW Clusters

Description

Initialize independent uniform random number streams to be used in a SNOW cluster. It uses either the L'Ecuyer's random number generator (package rlecuyer required) or the SPRNG generator (package rsprng required).

Usage

clusterSetupRNG (cl, type = "RNGstream", ...)

clusterSetupRNGstream (cl, seed=rep(12345,6), ...)
clusterSetupSPRNG (cl, seed = round(2^32 * runif(1)),
                   prngkind = "default", para = 0, ...) 

Arguments

cl
Cluster object.

type
type="RNGstream" (default) initializes the L'Ecuyer's RNG. type="SPRNG" initializes the SPRNG generator.

...
Arguments passed to the underlying function (see details bellow).

seed
Integer value (SPRNG) or a vector of six integer values (RNGstream) used as seed for the RNG.

prngkind
Character string naming generator type used with SPRNG.

para
Additional parameters for the generator.

Details

clusterSetupRNG calls (subject to its argument values) one of the other functions, passing arguments (cl, ...). If the "SPRNG" type is used, then the function clusterSetupSPRNG is called. If the "RNGstream" type is used, then the function clusterSetupRNGstream is called.

clusterSetupSPRNG loads the rsprng package and initializes separate streams on each node. For further details see the documentation of init.sprng. The generator on the master is not affected. NOTE: SPRNG is currently not supported.

clusterSetupRNGstream loads the rlecuyer package, creates one stream per node and distributes the stream states to the nodes.

For more details see http://www.stat.uiowa.edu/~luke/R/cluster/cluster.html.

Examples

  ## Not run: 
clusterSetupSPRNG(cl)
clusterSetupSPRNG(cl, seed=1234)
clusterSetupRNG(cl, seed=rep(1,6))
  
## End(Not run)

[Package snow version 0.4-3 Index]